+++ apps/backoffice-v2/CHANGELOG.md
@@ -1,11 +1,23 @@
 # @ballerine/backoffice-v2
 
+## 0.5.41
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.37
+  - @ballerine/workflow-browser-sdk@0.5.36
+  - @ballerine/workflow-node-sdk@0.5.36
+
 ## 0.5.40
 
 ### Patch Changes
 
 - Updated dependencies
   - @ballerine/ui@0.3.25
+  - @ballerine/common@0.7.36
+  - @ballerine/workflow-browser-sdk@0.5.35
+  - @ballerine/workflow-node-sdk@0.5.35
 
 ## 0.5.39