+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,19 @@
 # web-ui-sdk
 
+## 1.4.34
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.37
+
+## 1.4.33
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.36
+
 ## 1.4.32
 
 ### Patch Changes